refactor: split FeatureServer handlers and services (#228)
