+++ src/Honua.Server/Features/FeatureServer/FeatureServerQueryHandler.cs
@@ -14,23 +14,17 @@ using Honua.Core.Queries.Filters.Cql2;
 using Honua.Server.Features.FeatureServer.Models;
 using Honua.Server.Features.FeatureServer.Services;
 using Honua.Server.Features.Infrastructure.Models;
-using Honua.Server.Features.Infrastructure.Validation;
 
 namespace Honua.Server.Features.FeatureServer;
 
 /// <summary>
-/// Handler for FeatureServer operations that consolidates dependencies to reduce DI coupling.
-/// Addresses architectural limit of 5 dependencies per endpoint / 4 per handler.
+/// Handler for FeatureServer query operations.
 /// </summary>
-/// <remarks>
-/// Initializes a new FeatureServerHandler with required dependencies.
-/// Uses FeatureServerServices to aggregate query-related services, reducing dependency count to 4.
-/// </remarks>
-internal sealed class FeatureServerHandler(
+internal sealed class FeatureServerQueryHandler(
     ILayerCatalog layerCatalog,
     IFeatureStore featureStore,
-    FeatureServerServices services,
-    ILogger<FeatureServerHandler> logger)
+    IFeatureServerQueryServices queryServices,
+    ILogger<FeatureServerQueryHandler> logger)
 {
     private static readonly char[] _coordinateSeparators = { ',', ' ' };
     private static readonly HashSet<string> _allowedCoreOrderByFields = new(StringComparer.OrdinalIgnoreCase)
@@ -42,8 +36,8 @@ internal sealed class FeatureServerHandler(
     };
     private readonly ILayerCatalog _layerCatalog = layerCatalog ?? throw new ArgumentNullException(nameof(layerCatalog));
     private readonly IFeatureStore _featureStore = featureStore ?? throw new ArgumentNullException(nameof(featureStore));
-    private readonly FeatureServerServices _services = services ?? throw new ArgumentNullException(nameof(services));
-    private readonly ILogger<FeatureServerHandler> _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+    private readonly IFeatureServerQueryServices _queryServices = queryServices ?? throw new ArgumentNullException(nameof(queryServices));
+    private readonly ILogger<FeatureServerQueryHandler> _logger = logger ?? throw new ArgumentNullException(nameof(logger));
 
     /// <summary>
     /// Executes a feature query operation with proper validation and formatting.
@@ -83,7 +77,7 @@ internal sealed class FeatureServerHandler(
             }
 
             // Apply limits enforcement
-            QueryValidationResult validationResult = _services.QueryValidator.ValidateQueryLimits(queryParams);
+            QueryValidationResult validationResult = _queryServices.ValidateQueryLimits(queryParams);
             if (!validationResult.IsValid)
             {
                 return GeoServicesErrorHelpers.CreateBadRequestError(
@@ -107,7 +101,7 @@ internal sealed class FeatureServerHandler(
                     [geometryError ?? "Geometry parameter is invalid."]);
             }
 
-            var inputSrid = await ResolveSridAsync(validatedParams.InSr, parsedGeometry?.SpatialReference, cancellationToken);
+            var inputSrid = await _queryServices.ResolveSridAsync(validatedParams.InSr, parsedGeometry?.SpatialReference, cancellationToken);
             if (!string.IsNullOrWhiteSpace(validatedParams.InSr) && !inputSrid.HasValue)
             {
                 return GeoServicesErrorHelpers.CreateBadRequestError(
@@ -120,7 +114,7 @@ internal sealed class FeatureServerHandler(
                 inputSrid = layer.SpatialReference.Srid;
             }
 
-            var outputSrid = await ResolveSridAsync(validatedParams.OutSr, null, cancellationToken);
+            var outputSrid = await _queryServices.ResolveSridAsync(validatedParams.OutSr, null, cancellationToken);
             if (!string.IsNullOrWhiteSpace(validatedParams.OutSr) && !outputSrid.HasValue)
             {
                 return GeoServicesErrorHelpers.CreateBadRequestError(
@@ -146,7 +140,7 @@ internal sealed class FeatureServerHandler(
                 {
                     try
                     {
-                        sqlFilter = _services.SqlFilterTranslator.Translate(filterExpression, layer);
+                        sqlFilter = _queryServices.TranslateFilter(filterExpression, layer);
                     }
                     catch (ArgumentException)
                     {
@@ -224,7 +218,7 @@ internal sealed class FeatureServerHandler(
             string[]? outFields = string.IsNullOrEmpty(validatedParams.OutFields) ? null :
                 [.. validatedParams.OutFields.Split(',', StringSplitOptions.RemoveEmptyEntries).Select(f => f.Trim())];
 
-            (object? formattedResponse, string? contentType) = _services.QueryFormatter.FormatQueryResult(
+            (object? formattedResponse, string? contentType) = _queryServices.FormatQueryResult(
                 result,
                 layer,
                 validatedParams.F ?? "json",
@@ -309,7 +303,7 @@ internal sealed class FeatureServerHandler(
             var relationship = relationshipMaybe.Value;
 
             // Apply limits enforcement
-            RelatedRecordsValidationResult validationResult = _services.QueryValidator.ValidateRelatedRecordsLimits(queryParams);
+            RelatedRecordsValidationResult validationResult = _queryServices.ValidateRelatedRecordsLimits(queryParams);
             if (!validationResult.IsValid)
             {
                 return GeoServicesErrorHelpers.CreateBadRequestError(
@@ -373,558 +367,6 @@ internal sealed class FeatureServerHandler(
             return GeoServicesErrorHelpers.CreateInternalServerError("Related records query execution failed");
         }
     }
-
-    /// <summary>
-    /// Handles applyEdits requests for adding, updating, and deleting features.
-    /// </summary>
-    public async Task<IResult> HandleApplyEditsAsync(
-        string serviceId,
-        int layerId,
-        ApplyEditsRequest request,
-        Honua.Core.Configuration.EditLimits editLimits,
-        CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            FeatureServerLog.ApplyEditsRequested(_logger, serviceId, layerId,
-                request.Adds?.Length ?? 0,
-                request.Updates?.Length ?? 0,
-                request.Deletes?.Length ?? 0);
-
-            // Validate service and layer exist
-            var validationResult = await ValidateServiceAndLayer(serviceId, layerId, cancellationToken);
-            if (validationResult != null)
-            {
-                return validationResult;
-            }
-
-            var service = await _layerCatalog.GetServiceAsync(serviceId, cancellationToken);
-            var layer = service!.GetLayer(layerId)!;
-
-            // Validate edit limits
-            var limitsValidationResult = ValidateEditLimits(request, editLimits);
-            if (limitsValidationResult != null)
-            {
-                return limitsValidationResult;
-            }
-
-            var totalCount = (request.Adds?.Length ?? 0) + (request.Updates?.Length ?? 0) + (request.Deletes?.Length ?? 0);
-            if (totalCount == 0)
-            {
-                return Results.Json(new ApplyEditsResponse { Success = true },
-                    FeatureServerJsonContext.Default.ApplyEditsResponse,
-                    contentType: "application/json");
-            }
-
-            // Process edit operations
-            var editContext = ProcessEditOperations(request, layer);
-
-            // Handle validation errors with rollback if needed
-            if (editContext.HasValidationErrors && request.RollbackOnFailure)
-            {
-                return CreateRollbackResponse(editContext, serviceId, layerId);
-            }
-
-            // Execute edits in the database
-            var editResult = await ExecuteEdits(layer!.Id, editContext, request, cancellationToken);
-
-            // Build and return final response
-            return CreateFinalResponse(editContext, editResult, serviceId, layerId);
-        }
-        catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-        {
-            throw;
-        }
-        catch (Exception ex)
-        {
-            FeatureServerLog.ApplyEditsFailed(_logger, serviceId, layerId, ex.Message, ex);
-            return GeoServicesErrorHelpers.CreateInternalServerError("Apply edits failed");
-        }
-    }
-
-    /// <summary>
-    /// Validates that the service and layer exist
-    /// </summary>
-    private async Task<IResult?> ValidateServiceAndLayer(string serviceId, int layerId, CancellationToken cancellationToken)
-    {
-        var service = await _layerCatalog.GetServiceAsync(serviceId, cancellationToken);
-        if (service == null)
-        {
-            FeatureServerLog.ServiceNotFound(_logger, serviceId);
-            return GeoServicesErrorHelpers.CreateNotFoundError($"Service '{serviceId}' not found");
-        }
-
-        var layer = service.GetLayer(layerId);
-        if (layer == null)
-        {
-            FeatureServerLog.LayerNotFound(_logger, serviceId, layerId);
-            return GeoServicesErrorHelpers.CreateNotFoundError($"Layer {layerId} not found in service '{serviceId}'");
-        }
-
-        return null;
-    }
-
-    /// <summary>
-    /// Validates edit operation counts against configured limits
-    /// </summary>
-    private static IResult? ValidateEditLimits(ApplyEditsRequest request, Honua.Core.Configuration.EditLimits editLimits)
-    {
-        var addCount = request.Adds?.Length ?? 0;
-        var updateCount = request.Updates?.Length ?? 0;
-        var deleteCount = request.Deletes?.Length ?? 0;
-        var totalCount = addCount + updateCount + deleteCount;
-
-        if (addCount > editLimits.MaxFeaturesPerEdit ||
-            updateCount > editLimits.MaxFeaturesPerEdit ||
-            deleteCount > editLimits.MaxFeaturesPerEdit)
-        {
-            return GeoServicesErrorHelpers.CreateBadRequestError(
-                "Too many features in a single edit operation",
-                [$"Maximum per operation: {editLimits.MaxFeaturesPerEdit}"]);
-        }
-
-        if (totalCount > editLimits.MaxEditsPerTransaction)
-        {
-            return GeoServicesErrorHelpers.CreateBadRequestError(
-                "Too many edits in a single request",
-                [$"Maximum per request: {editLimits.MaxEditsPerTransaction}"]);
-        }
-
-        return null;
-    }
-
-    /// <summary>
-    /// Processes add, update, and delete operations from the request
-    /// </summary>
-    private EditOperationContext ProcessEditOperations(ApplyEditsRequest request, LayerDefinition layer)
-    {
-        var context = new EditOperationContext
-        {
-            AddResults = request.Adds is { Length: > 0 } ? new EditResult?[request.Adds.Length] : null,
-            UpdateResults = request.Updates is { Length: > 0 } ? new EditResult?[request.Updates.Length] : null,
-            DeleteResults = request.Deletes is { Length: > 0 } ? new EditResult?[request.Deletes.Length] : null
-        };
-
-        ProcessAddOperations(request, context, layer);
-        ProcessUpdateOperations(request, context, layer);
-        ProcessDeleteOperations(request, context);
-
-        return context;
-    }
-
-    /// <summary>
-    /// Processes add operations and tracks features to create
-    /// </summary>
-    private void ProcessAddOperations(ApplyEditsRequest request, EditOperationContext context, LayerDefinition layer)
-    {
-        if (request.Adds == null)
-            return;
-
-        for (var i = 0; i < request.Adds.Length; i++)
-        {
-            try
-            {
-                var newFeature = BuildFeatureFromGeoServices(request.Adds[i], 0, layer);
-                context.CreateFeatures.Add(newFeature);
-                context.CreateIndexes.Add(i);
-            }
-            catch (ArgumentException ex)
-            {
-                context.HasValidationErrors = true;
-                context.AddResults![i] = CreateFailureResult(
-                    code: 1000,
-                    description: ex.Message);
-            }
-            catch (Exception)
-            {
-                context.HasValidationErrors = true;
-                context.AddResults![i] = CreateFailureResult(
-                    code: 1000,
-                    description: "Failed to add feature");
-            }
-        }
-    }
-
-    /// <summary>
-    /// Processes update operations and tracks features to update
-    /// </summary>
-    private void ProcessUpdateOperations(ApplyEditsRequest request, EditOperationContext context, LayerDefinition layer)
-    {
-        if (request.Updates == null)
-            return;
-
-        for (var i = 0; i < request.Updates.Length; i++)
-        {
-            var update = request.Updates[i];
-            if (!TryGetObjectId(update.Attributes, out var objectId))
-            {
-                context.HasValidationErrors = true;
-                context.UpdateResults![i] = CreateFailureResult(
-                    code: 1001,
-                    description: "ObjectId is required for update operations");
-                continue;
-            }
-
-            try
-            {
-                var updateFeature = BuildFeatureFromGeoServices(update, objectId, layer);
-                context.UpdateFeatures.Add(updateFeature);
-                context.UpdateIndexes.Add(i);
-                context.UpdateObjectIds.Add(objectId);
-            }
-            catch (ArgumentException ex)
-            {
-                context.HasValidationErrors = true;
-                context.UpdateResults![i] = CreateFailureResult(
-                    code: 1002,
-                    description: ex.Message,
-                    objectId: objectId);
-            }
-            catch (Exception)
-            {
-                context.HasValidationErrors = true;
-                context.UpdateResults![i] = CreateFailureResult(
-                    code: 1002,
-                    description: "Failed to update feature",
-                    objectId: objectId);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Processes delete operations and tracks features to delete
-    /// </summary>
-    private static void ProcessDeleteOperations(ApplyEditsRequest request, EditOperationContext context)
-    {
-        if (request.Deletes == null)
-            return;
-
-        for (var i = 0; i < request.Deletes.Length; i++)
-        {
-            if (!TryConvertToLong(request.Deletes[i], out var objectId))
-            {
-                context.HasValidationErrors = true;
-                context.DeleteResults![i] = CreateFailureResult(
-                    code: 1003,
-                    description: "Invalid ObjectId for delete operation");
-                continue;
-            }
-
-            context.DeleteIds.Add(objectId);
-            context.DeleteIndexes.Add(i);
-        }
-    }
-
-    /// <summary>
-    /// Executes the validated edit operations in the database
-    /// </summary>
-    private async Task<FeatureEditResult> ExecuteEdits(
-        int layerId,
-        EditOperationContext context,
-        ApplyEditsRequest request,
-        CancellationToken cancellationToken)
-    {
-        if (context.CreateFeatures.Count == 0 && context.UpdateFeatures.Count == 0 && context.DeleteIds.Count == 0)
-        {
-            return FeatureEditResult.Success(0, 0, 0);
-        }
-
-        var editBatch = FeatureEditBatch.Create(
-            creates: context.CreateFeatures.ToImmutableArray(),
-            updates: context.UpdateFeatures.ToImmutableArray(),
-            deletes: context.DeleteIds.ToImmutableArray(),
-            rollbackOnFailure: request.RollbackOnFailure,
-            useGlobalIds: request.UseGlobalIds);
-
-        var editResult = await _featureStore.ApplyEditsAsync(layerId, editBatch, cancellationToken);
-
-        ApplyResults(context.AddResults, context.CreateIndexes, editResult.CreateResults);
-        ApplyResults(context.UpdateResults, context.UpdateIndexes, editResult.UpdateResults);
-        ApplyResults(context.DeleteResults, context.DeleteIndexes, editResult.DeleteResults);
-
-        return editResult;
-    }
-
-    /// <summary>
-    /// Creates a rollback response when validation errors occur
-    /// </summary>
-    private IResult CreateRollbackResponse(EditOperationContext context, string serviceId, int layerId)
-    {
-        var rollbackError = new EditError
-        {
-            Code = 1000,
-            Description = "Operation rolled back due to validation failure"
-        };
-
-        ApplyRollbackResults(context.AddResults, context.CreateIndexes, null, rollbackError);
-        ApplyRollbackResults(context.UpdateResults, context.UpdateIndexes, context.UpdateObjectIds, rollbackError);
-        ApplyRollbackResults(context.DeleteResults, context.DeleteIndexes, context.DeleteIds, rollbackError);
-
-        var response = new ApplyEditsResponse
-        {
-            AddResults = FinalizeResults(context.AddResults),
-            UpdateResults = FinalizeResults(context.UpdateResults),
-            DeleteResults = FinalizeResults(context.DeleteResults),
-            Success = false
-        };
-
-        FeatureServerLog.ApplyEditsCompleted(_logger, serviceId, layerId, false);
-
-        return Results.Json(response, FeatureServerJsonContext.Default.ApplyEditsResponse,
-            contentType: "application/json");
-    }
-
-    /// <summary>
-    /// Creates the final response after all operations complete
-    /// </summary>
-    private IResult CreateFinalResponse(EditOperationContext context, FeatureEditResult editResult, string serviceId, int layerId)
-    {
-        var finalAddResults = FinalizeResults(context.AddResults);
-        var finalUpdateResults = FinalizeResults(context.UpdateResults);
-        var finalDeleteResults = FinalizeResults(context.DeleteResults);
-        var allSuccess = AreAllResultsSuccessful(finalAddResults) &&
-                         AreAllResultsSuccessful(finalUpdateResults) &&
-                         AreAllResultsSuccessful(finalDeleteResults) &&
-                         !editResult.WasRolledBack &&
-                         !context.HasValidationErrors;
-
-        var finalResponse = new ApplyEditsResponse
-        {
-            AddResults = finalAddResults,
-            UpdateResults = finalUpdateResults,
-            DeleteResults = finalDeleteResults,
-            Success = allSuccess
-        };
-
-        FeatureServerLog.ApplyEditsCompleted(_logger, serviceId, layerId, allSuccess);
-
-        return Results.Json(finalResponse, FeatureServerJsonContext.Default.ApplyEditsResponse,
-            contentType: "application/json");
-    }
-
-    /// <summary>
-    /// Context object to track edit operations state
-    /// </summary>
-    private sealed class EditOperationContext
-    {
-        public EditResult?[]? AddResults { get; init; }
-        public EditResult?[]? UpdateResults { get; init; }
-        public EditResult?[]? DeleteResults { get; init; }
-        public List<Feature> CreateFeatures { get; } = new();
-        public List<int> CreateIndexes { get; } = new();
-        public List<Feature> UpdateFeatures { get; } = new();
-        public List<int> UpdateIndexes { get; } = new();
-        public List<long> UpdateObjectIds { get; } = new();
-        public List<long> DeleteIds { get; } = new();
-        public List<int> DeleteIndexes { get; } = new();
-        public bool HasValidationErrors { get; set; }
-    }
-
-    private Feature BuildFeatureFromGeoServices(GeoServicesFeature feature, long objectId, LayerDefinition layer)
-    {
-        byte[]? geometry = null;
-        if (feature.Geometry != null)
-        {
-            // Layer 1: Validate Esri JSON input
-            var esriValidation = _services.GeometryValidator.ValidateEsriJson(feature.Geometry);
-            if (!esriValidation.IsValid)
-            {
-                var errorMessages = string.Join("; ", esriValidation.Errors.Select(e => e.Message));
-                throw new ArgumentException($"Geometry validation failed: {errorMessages}");
-            }
-
-            var geometrySrid = feature.Geometry.SpatialReference?.Wkid
-                ?? feature.Geometry.SpatialReference?.LatestWkid
-                ?? layer.SpatialReference.Srid;
-            geometry = GeoServicesGeometryConverter.ConvertGeoServicesGeometryToWkb(feature.Geometry, geometrySrid);
-
-            // Layer 2: Validate WKB structure and size limits
-            var wkbValidation = _services.GeometryValidator.ValidateWkb(geometry);
-            if (!wkbValidation.IsValid)
-            {
-                var errorMessages = string.Join("; ", wkbValidation.Errors.Select(e => e.Message));
-                throw new ArgumentException($"WKB validation failed: {errorMessages}");
-            }
-        }
-
-        var attributesResult = layer.ValidateAttributes(
-            feature.Attributes,
-            ValidationExtensions.AttributeValidationMode.GeoServices);
-        if (!attributesResult.IsValid)
-        {
-            throw new ArgumentException(attributesResult.ErrorMessage ?? "Invalid attributes.");
-        }
-
-        return Feature.Create(objectId, geometry, attributesResult.Value!);
-    }
-
-    private static bool TryGetObjectId(Dictionary<string, object?>? attributes, out long objectId)
-    {
-        objectId = 0;
-
-        if (attributes == null || attributes.Count == 0)
-        {
-            return false;
-        }
-
-        foreach (var entry in attributes)
-        {
-            if (string.Equals(entry.Key, "objectid", StringComparison.OrdinalIgnoreCase))
-            {
-                return TryConvertToLong(entry.Value, out objectId);
-            }
-        }
-
-        return false;
-    }
-
-    private static bool TryConvertToLong(object? value, out long result)
-    {
-        result = 0;
-
-        if (value == null)
-        {
-            return false;
-        }
-
-        switch (value)
-        {
-            case long longValue:
-                result = longValue;
-                return true;
-            case int intValue:
-                result = intValue;
-                return true;
-            case short shortValue:
-                result = shortValue;
-                return true;
-            case byte byteValue:
-                result = byteValue;
-                return true;
-            case uint uintValue:
-                result = uintValue;
-                return true;
-            case ulong ulongValue when ulongValue <= long.MaxValue:
-                result = (long)ulongValue;
-                return true;
-            case string stringValue:
-                return long.TryParse(stringValue, NumberStyles.Integer, CultureInfo.InvariantCulture, out result);
-            case JsonElement element:
-                if (element.ValueKind == JsonValueKind.Number && element.TryGetInt64(out var elementLong))
-                {
-                    result = elementLong;
-                    return true;
-                }
-
-                if (element.ValueKind == JsonValueKind.String)
-                {
-                    return long.TryParse(element.GetString(), NumberStyles.Integer, CultureInfo.InvariantCulture, out result);
-                }
-
-                return false;
-            default:
-                return long.TryParse(Convert.ToString(value, CultureInfo.InvariantCulture), NumberStyles.Integer, CultureInfo.InvariantCulture, out result);
-        }
-    }
-
-    private static EditResult CreateFailureResult(int code, string description, long? objectId = null, string? globalId = null)
-    {
-        return new EditResult
-        {
-            ObjectId = objectId,
-            GlobalId = globalId,
-            Success = false,
-            Error = new EditError
-            {
-                Code = code,
-                Description = description
-            }
-        };
-    }
-
-    private static EditResult ConvertEditOperationResult(EditOperationResult result)
-    {
-        if (result.IsSuccess)
-        {
-            return new EditResult
-            {
-                ObjectId = result.ObjectId,
-                GlobalId = result.GlobalId,
-                Success = true
-            };
-        }
-
-        return CreateFailureResult(
-            result.ErrorCode,
-            result.ErrorMessage ?? "Operation failed",
-            result.ObjectId,
-            result.GlobalId);
-    }
-
-    private static void ApplyRollbackResults(EditResult?[]? results, List<int> indexes, List<long>? objectIds, EditError rollbackError)
-    {
-        if (results == null)
-        {
-            return;
-        }
-
-        for (var i = 0; i < indexes.Count; i++)
-        {
-            long? objectId = null;
-            if (objectIds != null && i < objectIds.Count)
-            {
-                objectId = objectIds[i];
-            }
-
-            results[indexes[i]] = CreateFailureResult(rollbackError.Code, rollbackError.Description, objectId);
-        }
-    }
-
-    private static void ApplyResults(EditResult?[]? results, List<int> indexes, ImmutableArray<EditOperationResult> operationResults)
-    {
-        if (results == null)
-        {
-            return;
-        }
-
-        var count = Math.Min(indexes.Count, operationResults.Length);
-        for (var i = 0; i < count; i++)
-        {
-            results[indexes[i]] = ConvertEditOperationResult(operationResults[i]);
-        }
-
-        for (var i = count; i < indexes.Count; i++)
-        {
-            results[indexes[i]] ??= CreateFailureResult(1000, "Operation failed");
-        }
-    }
-
-    private static EditResult[]? FinalizeResults(EditResult?[]? results)
-    {
-        if (results == null)
-        {
-            return null;
-        }
-
-        var finalized = new EditResult[results.Length];
-        for (var i = 0; i < results.Length; i++)
-        {
-            finalized[i] = results[i] ?? CreateFailureResult(1000, "Operation failed");
-        }
-
-        return finalized;
-    }
-
-    private static bool AreAllResultsSuccessful(EditResult[]? results)
-    {
-        if (results == null)
-        {
-            return true;
-        }
-
-        return results.All(result => result.Success);
-    }
-
 
     /// <summary>
     /// Builds a FeatureQuery from query parameters
@@ -1184,151 +626,6 @@ internal sealed class FeatureServerHandler(
         };
     }
 
-    private async Task<int?> ResolveSridAsync(
-        string? srValue,
-        GeoServicesSpatialReference? geometrySpatialReference,
-        CancellationToken cancellationToken)
-    {
-        var srid = await ParseSridAsync(srValue, cancellationToken);
-        if (srid.HasValue)
-        {
-            return srid;
-        }
-
-        if (geometrySpatialReference != null)
-        {
-            if (geometrySpatialReference.Wkid > 0)
-            {
-                return geometrySpatialReference.Wkid;
-            }
-
-            if (geometrySpatialReference.LatestWkid.HasValue)
-            {
-                return geometrySpatialReference.LatestWkid.Value;
-            }
-
-            if (!string.IsNullOrWhiteSpace(geometrySpatialReference.Wkt))
-            {
-                return await _services.CrsDetectionService.DetectFromWktAsync(geometrySpatialReference.Wkt);
-            }
-        }
-
-        return null;
-    }
-
-    private async Task<int?> ParseSridAsync(string? srValue, CancellationToken cancellationToken)
-    {
-        if (string.IsNullOrWhiteSpace(srValue))
-        {
-            return null;
-        }
-
-        var trimmed = srValue.Trim();
-
-        if (int.TryParse(trimmed, NumberStyles.Integer, CultureInfo.InvariantCulture, out var srid))
-        {
-            return srid;
-        }
-
-        if (trimmed.StartsWith('{'))
-        {
-            if (TryParseSpatialReferenceJson(trimmed, out var wkid, out var wkt, out var name))
-            {
-                if (wkid.HasValue)
-                {
-                    return wkid.Value;
-                }
-
-                if (!string.IsNullOrWhiteSpace(name))
-                {
-                    var epsg = _services.CrsDetectionService.DetectFromEpsgCode(name);
-                    if (epsg.HasValue)
-                    {
-                        return epsg.Value;
-                    }
-                }
-
-                if (!string.IsNullOrWhiteSpace(wkt))
-                {
-                    return await _services.CrsDetectionService.DetectFromWktAsync(wkt);
-                }
-            }
-        }
-
-        var detected = _services.CrsDetectionService.DetectFromEpsgCode(trimmed);
-        if (detected.HasValue)
-        {
-            return detected.Value;
-        }
-
-        if (LooksLikeWkt(trimmed))
-        {
-            return await _services.CrsDetectionService.DetectFromWktAsync(trimmed);
-        }
-
-        return null;
-    }
-
-    private static bool LooksLikeWkt(string value)
-    {
-        return value.StartsWith("GEOGCS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("PROJCS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("GEOGCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("PROJCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("GEODCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("GEODETICCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("COMPD_CS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("COMPOUNDCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("VERT_CS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("VERTCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("LOCAL_CS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("LOCALCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("BOUNDCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("ENGCRS[", StringComparison.OrdinalIgnoreCase)
-               || value.StartsWith("ENGINEERINGCRS[", StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static bool TryParseSpatialReferenceJson(string value, out int? wkid, out string? wkt, out string? name)
-    {
-        wkid = null;
-        wkt = null;
-        name = null;
-
-        try
-        {
-            using var document = JsonDocument.Parse(value);
-            var root = document.RootElement;
-
-            if (root.TryGetProperty("wkid", out var wkidElement) && wkidElement.TryGetInt32(out var wkidValue))
-            {
-                wkid = wkidValue;
-            }
-
-            if (!wkid.HasValue &&
-                root.TryGetProperty("latestWkid", out var latestElement) &&
-                latestElement.TryGetInt32(out var latestWkid))
-            {
-                wkid = latestWkid;
-            }
-
-            if (root.TryGetProperty("wkt", out var wktElement))
-            {
-                wkt = wktElement.GetString();
-            }
-
-            if (root.TryGetProperty("name", out var nameElement))
-            {
-                name = nameElement.GetString();
-            }
-
-            return wkid.HasValue || !string.IsNullOrWhiteSpace(wkt) || !string.IsNullOrWhiteSpace(name);
-        }
-        catch (JsonException)
-        {
-            return false;
-        }
-    }
-
     private static bool TryParseGeoServicesGeometry(
         string? geometryText,
         string? geometryType,
@@ -1520,7 +817,7 @@ internal sealed class FeatureServerHandler(
         foreach (var feature in result.Items)
         {
             if (feature.Attributes?.TryGetValue(relationship.DestinationForeignKeyField, out object? fkValue) == true &&
-                TryConvertToLong(fkValue, out var originId))
+                FeatureServerValueParser.TryConvertToLong(fkValue, out var originId))
             {
                 if (!featuresByOriginId.TryGetValue(originId, out var bucket))
                 {
