+++ src/Honua.Server/Features/FeatureServer/Services/FeatureServerQueryServices.cs
@@ -1,0 +1,46 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+using Honua.Core.Features.Catalog.Domain;
+using Honua.Core.Features.FeatureStore.Domain;
+using Honua.Core.Queries.Filters;
+using Honua.Server.Features.FeatureServer.Models;
+
+namespace Honua.Server.Features.FeatureServer.Services;
+
+internal sealed class FeatureServerQueryServices(
+    IQueryFormatter queryFormatter,
+    IFeatureQueryValidator queryValidator,
+    ISqlFilterTranslator sqlFilterTranslator,
+    SpatialReferenceResolver spatialReferenceResolver) : IFeatureServerQueryServices
+{
+    private readonly IQueryFormatter _queryFormatter = queryFormatter ?? throw new ArgumentNullException(nameof(queryFormatter));
+    private readonly IFeatureQueryValidator _queryValidator = queryValidator ?? throw new ArgumentNullException(nameof(queryValidator));
+    private readonly ISqlFilterTranslator _sqlFilterTranslator = sqlFilterTranslator ?? throw new ArgumentNullException(nameof(sqlFilterTranslator));
+    private readonly SpatialReferenceResolver _spatialReferenceResolver =
+        spatialReferenceResolver ?? throw new ArgumentNullException(nameof(spatialReferenceResolver));
+
+    public QueryValidationResult ValidateQueryLimits(QueryParameters queryParams)
+        => _queryValidator.ValidateQueryLimits(queryParams);
+
+    public RelatedRecordsValidationResult ValidateRelatedRecordsLimits(QueryRelatedRecordsParameters queryParams)
+        => _queryValidator.ValidateRelatedRecordsLimits(queryParams);
+
+    public SqlFragment? TranslateFilter(FilterExpression filterExpression, LayerDefinition layer)
+        => _sqlFilterTranslator.Translate(filterExpression, layer);
+
+    public Task<int?> ResolveSridAsync(
+        string? srValue,
+        GeoServicesSpatialReference? geometrySpatialReference,
+        CancellationToken cancellationToken = default)
+        => _spatialReferenceResolver.ResolveSridAsync(srValue, geometrySpatialReference, cancellationToken);
+
+    public (object? Response, string? ContentType) FormatQueryResult(
+        QueryResult<Feature> result,
+        LayerDefinition layer,
+        string format,
+        bool returnGeometry,
+        int? outputSrid,
+        string[]? outFields)
+        => _queryFormatter.FormatQueryResult(result, layer, format, returnGeometry, outputSrid, outFields);
+}
