+++ src/Honua.Server/Features/FeatureServer/Services/IFeatureServerGeometryServices.cs
@@ -1,0 +1,14 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+using Honua.Core.Features.Geometry.Domain;
+using Honua.Server.Features.FeatureServer.Models;
+
+namespace Honua.Server.Features.FeatureServer.Services;
+
+internal interface IFeatureServerGeometryServices
+{
+    GeometryValidationResult ValidateEsriJson(GeoServicesGeometry? geometry);
+
+    GeometryValidationResult ValidateWkb(byte[]? wkb);
+}
