+++ src/Honua.Server/Features/FeatureServer/Services/IFeatureServerQueryServices.cs
@@ -1,0 +1,31 @@
+// Copyright (c) Honua. All rights reserved.
+// Licensed under the Elastic License 2.0. See LICENSE in the project root.
+
+using Honua.Core.Features.Catalog.Domain;
+using Honua.Core.Features.FeatureStore.Domain;
+using Honua.Core.Queries.Filters;
+using Honua.Server.Features.FeatureServer.Models;
+
+namespace Honua.Server.Features.FeatureServer.Services;
+
+internal interface IFeatureServerQueryServices
+{
+    QueryValidationResult ValidateQueryLimits(QueryParameters queryParams);
+
+    RelatedRecordsValidationResult ValidateRelatedRecordsLimits(QueryRelatedRecordsParameters queryParams);
+
+    SqlFragment? TranslateFilter(FilterExpression filterExpression, LayerDefinition layer);
+
+    Task<int?> ResolveSridAsync(
+        string? srValue,
+        GeoServicesSpatialReference? geometrySpatialReference,
+        CancellationToken cancellationToken = default);
+
+    (object? Response, string? ContentType) FormatQueryResult(
+        QueryResult<Feature> result,
+        LayerDefinition layer,
+        string format,
+        bool returnGeometry,
+        int? outputSrid,
+        string[]? outFields);
+}
